VS Codebereinigung

+++ UI/CustomControls/EquationLabel.cs
@@ -84,8 +84,8 @@ namespace BauphysikToolWPF.UI.CustomControls
             return (double)element.GetValue(SubscriptFontSizeProperty);
         }
 
-        public static readonly DependencyProperty SubscriptFontSizeProperty = DependencyProperty.Register("SubscriptFontSize", typeof(double), typeof(EquationLabel), new PropertyMetadata((double)12.0));
-        
+        public static readonly DependencyProperty SubscriptFontSizeProperty = DependencyProperty.Register("SubscriptFontSize", typeof(double), typeof(EquationLabel), new PropertyMetadata(12.0));
+
         public static void SetFontSize(UIElement element, double value)
         {
             element.SetValue(FontSizeProperty, value);
@@ -95,7 +95,7 @@ namespace BauphysikToolWPF.UI.CustomControls
             return (double)element.GetValue(FontSizeProperty);
         }
 
-        public static new readonly DependencyProperty FontSizeProperty = DependencyProperty.Register("FontSize", typeof(double), typeof(EquationLabel), new PropertyMetadata((double)14.0));
+        public static new readonly DependencyProperty FontSizeProperty = DependencyProperty.Register("FontSize", typeof(double), typeof(EquationLabel), new PropertyMetadata(14.0));
 
     }
 }
